Fix the tests and remove some residual cruft from them
The tests were not in line with the latest updates to the module, causing some
to fail legitimately and some to fail due to incorrect setup.
Also 2 of the assertions were basically useless in the current setup, and some
variables were defined in the wrong place or defined and never use.