Remove token-url management from template
Missing token-url is purely a developer error and has nothing to do in the template.
Failure to retrieve the token should be handled on a case by case basis depending
on what kind of error the server could return, so it should probably be delegated to
a callback from the consumer, or simply throw if nothing is provided.