Remove token-url provider entirely and pass it from the directive attrs instead
The token-url provider holds as a prerequisite that the token retrieval URL is
a specific route in the API and does not change (or else we would end up with code
entanglement and race conditions if several img-uploads are used on the same view).
To support multiple token retrieval URLs, it makes more sense to make the token-url
an attribute of the directive so that a different URL can be passed when we create
the directive.