Check and clean the form data that is sent to Amazon
There are precise requirements on the data that must be sent to Amazon,
especially as it needs to be exactly identical to the fields that can
be retrieved in the signed policy.
Add a stricter way to assign those fields and check that each and
every one of them is present so we can minimize the Amazon XML
errors we get.