Update the tests with the new organization and policy token formatting
The tests were still spawning a controller and testing against it instead
of compiling the directive itself. Which is incompatible with using
required objects bound to the directive's controller.
Also update the tests to reflect the new format of the policy token,
corresponding to the exact field names amazon expects.