Trigger the callback on successful or failing completion of upload
The consumer might need to be able to perform certain actions when the
upload is over, to recover from an error or keep moving after a
success.
In both cases, try and transmit as much data as we can, since these
callbacks can be called on multiple items in the same instance.