Remove imgUpload as an all-in directive with a template
The initial approach of using imgUpload as a light wrapper around
angular-file-upload did not produce up-to-par results as it needs
to be constantly tweaked until we reproduce the entire angular-file-upload
API which is not a desirable outcome.
Moreover, the template needs to be uniquely tweaked for each instance so
it's not practical to make this an element directive.